requirements.txt
python-telegram-bot==20.6
pymongo==4.5.0
prometheus-client==0.17.1
tenacity==8.2.3
pyrate-limiter==2.10.0
python-dotenv==1.0.0
motor==3.3.1

# Dependências de teste
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0 